Update the .copr/Makefile and how it makes SRPM files.
The target was relying on running autogen.sh and then running configure
to run make srpm.  The problem here is the system that copr runs this
Makefile on.  Right now, that's F28 and not all the build deps to run
configure are present.  So adjust this to make an SRPM file directly,
but that also means that the source has to be autoconf'ed from the spec
file.