Return an empty list instead of NULL for no symbols.
In get_elf_symbol_list, return an empty list instead of NULL for Elf
objects that do not contain the expected section (.dynsym or .symtab).
Objects missing .dynsym are especially common, since statically linked
binaries don't need it. In either case, there are no symbols to list, so
return an empty list.

Failing to get a data pointer for a section is an actual error, so
continue to return NULL for those cases.