Fix hardlink extraction in extract_rpm()
Since we are extracting files to a temporary location, libarchive needs
to know the hardlink destination in the right path.  The only way to
determine if an archive_entry is a hard link is to look at the nlink
value, which is > 1 for hard links (assuming the file type is S_ISREG).