Simplify the coveralls upload
The check-code-coverage target already generates the .info file that
coveralls needs, so we don't need to run gcov again or worry about
uploading to coveralls from inside of the test-env container.

cpp-coveralls claims to support lcov info files, but it bombs out on the
test name lines. Switch to lcoveralls, available via rubygems.