Fix the test suite runners.
CU_basic_run_tests doesn't return set or return a failure code failed
tests. A separate function call is needed to find out if the test suite
succeeded or not. All of the tests were returning success whether the
actual test assertions succeeded or not.

Move the test initialization, running, failure checking, and cleanup
into a separate shared source file so that if this turns out to be wrong
too it only needs to be fixed in one place. Convert the main() functions
in the rest of the test-*.c files to get_suite(), which just creates the
CU_pSuite, adds the test to it, and returns the suite object.