Add a valgrind check
Add ax_valgrind_check.m4 from autoconf-archive, which is licensed under
the FSF All Permissive License. This adds a check-valgrind target, which
runs tests under valgrind.

Add the extra configure.ac and Makefile.am bits, and run check-valgrind
in .travis.yml. Rearrange the travis script so that it stops at the
first failure, to make the output easier to read.