Add the 'desktop' inspection to librpminspect
Or at least the beginnings of it.  What this inspection does is scan all
subpackages in a build for desktop entry files.  Those files that end
with *.desktop and sometimes even *.directory.  It then does this:

1) Runs desktop-file-validate on the file.  If we get non-zero back, we
treat that as RESULT_BAD.  A lot of times we get stuff back on stdout
but otherwise the file is valid.  We report that stuff as RESULT_INFO.

2) For each desktop entry file, read it line by line looking for Exec=
and Icon= lines.

    a) For Exec= lines, look for the referenced executable in any
       subpackage in the build.  If not found, report that.  If found,
       make sure it's executable by everyone, otherwise report that.
       Executables can be explicitly referenced.  If not, look for them
       under /usr/bin, unless the desktop file is in a kde4
       subdirectory.  If it is, look for the executable in
       /usr/libexec/kde4.  We might be able to drop the KDE4 stuff
       entirely.

    b) For Icon= lines, look for the referenced icon file in any
       subpackage in the build.  Missing icons are reported.  Icons
       can be referenced by explicit path or relative to
       /usr/share/pixmaps.  If an icon file lacks an extension, treat
       it as a theme icon and skip the check.

I'll likely revise this code tomorrow.  Probably can turn a number of
things in to functions.