Make sure test-inspect_elf.c opens helper files correctly.
The tests create a number of compiled ELF executables to use during unit
tests.  When the unit tests open those files, make sure they are using
the correct path to them.  If you build in the source directory, that's
just the current working directory.  But if you use a separate build
directory, that needs to be accounted for.

When compiling the helper programs, set a macro in CFLAGS called
_BUILDDIR_ and set it to the value of $(abs_builddir).  In
test-inspect_elf.c, use this macro in the open(2) calls and just let C
do string concatenation for constant strings.