Allow the -T option in rpminspect(1) to work with negated values.
By default all inspections are run.  Or you can specify a list of
inspections to run separated by commas.  You might want to run "all but"
a handful of inspections.  The syntax for that is ! at the beginning.
For example:

    rpminspect -T manpage,desktop

Runs only the manpage and desktop inspections.  But:

    rpminspect -T !manpage,!desktop

Runs all inspections except manpage and desktop.  Figured this could be
useful for specifying per-package settings.

Note that when you use the first form (manpage,desktop), the assumption
is all tests are off by default and you are explicitly enabling the
tests you want.  If you use the negation syntax, then it changes to all
tests on by default except the ones you specify negated.  It is ok to
list non-negated tests with negation syntax, but it has no effect.  That
is, "!manpage,!desktop" is functionally the same as
"ALL,!manpage,!desktop".