Add hours and minutes to the timestamp in GITDATE
For automated builds, we may have multiple builds per day but the commit
hash would be different.  The hashes are not sortable, which would mean
dnf may or may not detect an upgrade available with the copr repo
updates.  To get sortable release values, extend the date stamp to
include hours and minutes in addition to year, month, and day.