Remove special handling for the before and after SRPM packages.
Let these be handled through the add_peer() code path.  This allows the
SRPM packages to get unpacked just like the binary packages and set up
peers the same way.  This is required for inspections that will look at
the contents of the source packages, such as patches and the spec file.