Drop log file check from is_local_build()
This was an added check to ensure a specified local directory is a koji
build.  But we're not going to do anything with the log files and for
some types of koji builds we may get different logs, so just assume the
user specified a usable directory that is a build and run with it.