Download and sort modulemd*txt files for KOJI_BUILD_MODULE types.
If we are downloading modules for inspection, make sure to collect the
modulemd YAML files and sort them out in to the appropriate
subdirectories as we lay them out.  For modules, there is a top level
modulemd.txt that should be in the after/ or before/ subdirectory.  Then
there are modulemd.ARCH.txt files in each subdirectory under after/ and
before/, including 'src' but *not* 'noarch'.  I have not found an XMLRPC
call that will tell me the architectures that apply to a particular
module, so I modified the loop to just try every architecture we come
across and assuming libcurl tells us CURLE_OK, assume we have it.  Skip
subsequent downloads for ones we already fetched (except I do that with
a stat now and it could probably be a boolean in the code....eh, another
commit).