rpminspect grew a dependency:  libyaml
Why?  Modules in the context of Koji and RPM builds in the world of
Fedora and Red Hat are collections of koji builds with additional
metadata in YAML format.  rpminspect reads this metadata to filter out
RPMs during module fetching.  It may do more in the future with the
metadata (maybe even inspections), but for now this is all it uses
modulemd.txt for.

Read more about modules here:
    https://docs.fedoraproject.org/en-US/modularity/

Why not libmodulemd?  I looked at this library and it's tied up with
glib and other libraries that rpminspect is not using.  To minimize
dependencies, I decided to just use libyaml directly.