Type corrections for Koji communcation.
epoch changed from a 'char *' to an 'int'.  module_service_build_id
changed from an int to a 'char *'.  I'm not sure why this happened, but
it appears to have happened in both Fedora and RHEL.  I suspect this is
going to be a source of chasing problems as the API expands.  That might
be a thing where the Koji communication needs to become a separate
library.