Bug fix in extract_rpm(): can't use headerIsEntry() on RPMTAG_FILENAMES
The RPMTAG_FILENAMES tag is a header extension in librpm, which means it
does not return from headerIsEntry().  We just have to call headerGet()
and if it doesn't return anything, we can deduce the payload is empty.
What I don't like about this is the inability to get any I/O erroring
reporting.