Add product_release member to struct rpminspect.
The product_release is determined by the builds you specify, but can be
specified on the command line in case rpminspect cannot determine it.
What rpminspect tries to do is grab a dist tag from the end of the NVR
and use that.  In cases where you are comparing builds from two
different products, you should specify the product release on the
command line.  rpminspect will only work if both builds have matching
dist tags.

What is a dist tag?  This is part of Fedora packaging policy.  A package
Release tag will look something like this:

    Release: 1%{?dist}

The %{?dist} expands to a string if set, but otherwise is empty.  In the
Fedora build system this is expanded to something like '.fc30' for
Fedora 30.  rpminspect just uses this string as-is since it tends to
vary by vendor.