rpminspect now has the -r/--release option for product release.
rpminspect will first try to determine the product release of the builds
specified by looking for a dist tag at the end of the NVR string.  If
those match between the before and after builds, it continues.  If they
do not match, it stops.  The command line option lets you specify the
product release manually.

Keep in mind that the product release will be used in data file tables
in rpminspect-data, so that's why I went with the dist tag since that is
already know and well established.