Add code to librpminspect to support the [javabytecode] block.
This commit adds a hash table to struct rpminspect which maps product
release strings to JVM major version numbers.  It expects the config
file to have a block like this:

    [javabytecode]
    fc29 = 53
    fc30 = 53
    default = 44

The default entry will be used if it can't find a key match to your
product release string.  If default is missing, rpminspect will fail.
You can avoid this by providing a default entry -or- disabling the
javabytecode test at runtime.