In inspect_javabytecode, fill out more of the tests.
Now that rpminspect has a javabytecode table in the configuration file
and that's read in to a hash table from init.c, the javabytecode test
can collect the expected JVM major version for the product release we
are testing against.  With that information, we can compare it against
the found JVM major versions in class files and report results.

The next parts of this test are extending the major version comparison
code to the jar_walker() routine to check class files inside jar files.
And also adding in the before/after JVM version comparison to report
changes to the maintainer.