Move JVM version checking to check_class_file()
This is a static file called by either the main driver of the inspection
or the jar_walker().  The main driver handles any .class files that
appear in the package payload while the jar_walker() handles further
diving in to unpacked .jar files found in the package.  The tests
performed end up being the same, so move them to a shared function.