Extend the javabytecode test to compare before/after JVM versions.
There is one last thing to complete in this test and that's checking to
see if the JVM major version changed from the before to after build.  If
it did, report it as a failure.

This commit only does that for the part of the inspection that runs
through all of the files in the package.  The next commit will extend
this to the .jar walker.