Hard links extract but return !ARCHIVE_OK from archive_read_extract() (#19)
The hard link is extracting fine, but we just don't get ARCHIVE_OK from
archive_read_extract().  This is filtered up from
archive_write_disk_posix.c which knows something happened, but it falls
through to the catchall error block.  I actually don't think an error
occurred because the extraction worked.  The correct place to fix this
is upstream, so I will be working on a fix there.  For now we can ignore
the return value if we are unpacking a hard link.  Not great since a
failure here is actually ignored, but I need a workaround for now.