Trim trailing slashes from release strings.
In case the user is specifying builds from local directories, make sure
to trim trailing slashes since they are not part of the dist tag.  This
is a convenience thing as it's common to hit tab to complete paths and
that will give you a trailing slash on directory names.