Support ~ expansion for --workdir=PATH specifications (#22)
~ expansion is generally handled by your shell, but when using long
options most shells do not expand the ~ character and pass the string
directly in the argument vector.

What I have done is add ~ expanding for the workdir argument if you have
a ~ in the string. See the glob(3) man page for details on the
functionality. I am using the GLOB_TILDE_CHECK functionality, so there
are some additional checks to ensure the user is there and such.

Since I am relying on glob(3), I will get GLOB_NOMATCH back if the path
does not exist. That is a side effect of using glob(3). So if you are
going to use the long option syntax of rpminspect, keep this in mind.