Add support for specifying architecture list on the command line (#27)
This is part 1 of a 2 part enhancement.  This commit adds code that
processes a new command line argument called -a or --arches.  You
specify a list of architectures you want to run rpminspect for,
delimited by commas.  The default mode of operation is to operate on all
found architectures, but there may be instances where you want to
restrict the run to one or a subset of architectures.

The syntax is:

    rpminspect -a ARCH,ARCH,ARCH BEFORE_BUILD AFTER_BUILD

There is no "-a ALL" option since the default is to operate on
everything.  Architecture names are checked against what the Koji hub
says it can build for.  In Fedora right now, a valid -a option would
look like this:

    rpminspect -a x86_64,aarch64,ppc64le

If you specify the -a option, the SRPM is not automatically included.
You need to add 'src' to the list.  Like this:

    rpminspect -a x86_64,aarch64,ppc64le,src

If you specify something like this in Fedora now:

    rpminspect -a hppa,mips,sparc64,src

You will get an error telling you hppa is an unsupported architecture.
rpminspect will terminate on the first invalid architecture specified.

The next commit will hook up the build gathering to the architecture
list.