Split the -T option functionality in to -T and -E (#28)
The purpose of this change is to avoid the \!testname syntax to exclude
a test from an rpminspect run.  By default, all tests that apply will
run.  There are now the -T and -E options:

If you specify -T, then all tests are disabled by default and only the
tests you specify on this option are enabled.  You can pass "-T ALL" to
enable all tests, but that is the default behavior.

If you specify -E, then all tests are enabled by default and only the
tests you specify on this option are disabled.  You can pass "-E ALL" to
disable all tests, but that is sort of pointless because rpminspect ends
up doing nothing.

The options cannot be used together.  The man page and README have been
updated.