Add the 'changedfiles' inspection.
This is one of the main functions of rpminspect and part of a set of
inspections coming.  The first is changedfiles and here's what it does:

It only runs when you are comparing two builds since it is not possible
to determine if a single build failed or not.  This inspection iterates
over each package and compares each file peer pair with the following
conditions:

    * It skips non-regular files.
    * It skips any files lacking a peer.  There are separate inspections
      to handle new, removed, and moved files.
    * It skips Java class files and jar files (these are handled by
      other inspections.

With those conditions met, it does this:

    * If the file is gzipped, it zcmps the peers and reports.
    * If the file is bzip2'ed, it bzcmps the peers and reports.
    * If the file is xz'ed, it xzcmps the peers and reports.
    * If the file is an ELF object, it runs eu-elfcmp on the peers and
      reports.
    * If the file is a gettext message catalog, it unpacks the catalogs
      to temporary files and runs 'diff -u' on the output and then
      reports.
    * If the files are public C or C++ header files, it preprocesses the
      files to remove comments, then runs 'diff -uw' on the output and
      reports.
    * If none of the special cases match, the SHA-256 digest of the
      peers are computed and compared.

For any of theses, a non-successful return of the comparing program or
non-matching digests results in a RESULT_VERIFY message along with the
output of the program (if applicable).

I am interested in feedback on this inspection as well as improvements
to the messages it outputs.  And bug reports too.  Please file as github
issues.