Improve the -f/--fetch-only mode.
When using rpminspect to download Koji builds, the program would just
put them in the working directory but the path name would be what
rpminspect would have been using internally.  This was ok, but a little
inconvenient if you are using rpminspect to quickly fetch a number of
NVRs.

The -f option is restricted to remote builds only and only a single
build specification on the command line.  The output will go to the NVR
subdirectory of the working directory.  By default the working directory
is /var/tmp/rpminspect.  If you use rpminspect to fetch
xpdf-4.01-1.fc31, it would go to:

    /var/tmp/rpminspect/xpdf-4.01-1.fc31

But you can specify the -w/--workdir option to give it a different path,
such as ~/downloaded-builds:

    rpminspect -v -w ~/downloaded-builds -f xpdf-4.01-1.fc31

And the xpdf build will be in ~/downloaded-builds.  The -a option works
as well for restricting architectures that are downloaded.