Start integration test suite in the tests/ subdirectory.
Functionality for the end user program will need come with new tests or
modifications to existing tests.  A test should be a script (do not care
about the language, but not compiled unless you are doing unit tests)
that takes environment variables like RPMINSPECT.  Output any useful
messages you want to output and return 0 on success and non-zero on
failure.