Fix config file detection
When a configuration file is passed on the command line, it should
always be used. In particular, that should mean erring out and telling
the user that their parameter was bad. When no config parameter was
passed, we should use the default if it exists. However, since
rpminspect is a new package (and choice of default rpminspect-data-*
depends on the distro), we should inform the user what they're missing.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>