Expand struct rpminspect with 'stat_whitelist_dir'
And add a new config file section called [vendor-data].  This begins the
settings to define where to find the vendor-specific files that are per
product release.  Starting with the stat-whitelist files, there is now a
setting for the stat-whitelist directory.  The default location is in
constants.h and duplicated in the example rpminspect.conf file, but the
vendor can change this to whatever path they want.

Pay attention to the comments in the rpminspect.conf file.  Settings
like this will take the form of *_dir because they are specifying a
directory location.  In the case of the stat-whitelist, there is a
different one per product release.  The actual files in that directory
should match your approved values for the product release.  As an
example, for Fedora Linux the product release values are the %{disttag}
values.