Add stat_whitelist to struct rpminspect.
The stat_whitelist is the contents of the stat-whitelist file for the
product release the program is running for.  It is populated by
init_stat_whitelist().  This population is done once for the lifetime of
the program so it is safe to call init_stat_whitelist() and act based on
its return value with true meaning it's either populated or I populated
it.  False indicates a missing whitelist or error while reading it.