Properly free the stat_whitelist, fix init_stat_whitelist()
Make sure the stat_whitelist is freed if it was ever used during
runtime.

When initializing the stat_whitelist, make sure the filenames do not
have leading slashes.  If we trim and get to the end of the string,
report that as an invalid line in the stat-whitelist file and then skip
it and go to the next line.  Should help reveal typos in those files
while still using the valid lines.