New release.sh script, remove final Makefile.am
The top level release.sh script is still a work in progress.  This
removes the final Makefile.am file.  The intent of the release.sh script
is for use in making new versioned releases with a single command.
Ultimately this should key on the success of Travis-CI output and then
automatically make a new release complete with a release entry on github
and uploaded artifacts.  Very close to that.