Add the 'upstream' inspection to librpminspect.
This is a new inspection for builds.  It only runs when you are
comparing a before and after build and it also only applies to SRPM
files.  Here's what it does:

1) It finds the SRPM peers for the before and after builds.
2) It gets the Version and Epoch from the RPM headers.
3) If the before and after Versions differ -OR- if the Epoch exists in
   either package and is different, then discontinue the inspection.
4) If the Versions and Epoch values are the same, do this:
       a) Report any SourceN files in the SRPM that appeared in the
          after build but were not in the before build.  These are
          flagged as RESULT_VERIFY.
       b) Report any SourceN files in the SRPM that changed content
          between the before and after build, mark as RESULT_BAD.  The
          idea here is catching a package rebuild of the same version
          but the source files changed.  This may be deliberate, but
          it could also be unexpected.
       c) Report any SourceN files in the SRPM that disappeared in the
          after build but were present in the before build.  These are
          marked as RESULT_VERIFY.

Content changes are done by computing the SHA-256 digest of file peers.
Care is taken to only examine those files that are SourceN files in the
spec file.  That list is obtained from the RPMTAG_SOURCE array in the
SRPM header.  This avoids examining patches and other SRPM members in
this inspection.