Discontinue use of cpp(1) in changedfiles.
The changedfiles inspection was running C and C++ headers through cpp in
order to remove comments, but this was behaving differently depending on
the system you were running on.  The problem is cpp is expanding macros
according to header files on the host system.  We don't want that, we
only want to remove comments.  There's no clean way to do this and since
it's really just a nice-to-have, I decided to remove it because this
still just gets raised as a RESULT_VERIFY for the maintainer to check.
Comments can change in the public header files and that's worth a
once-over in addition to prototypes and definitions changing.