In the 'metadata' inspection, do not assume the Vendor tag is set. (#43)
The Vendor tag may or may not be set depending on where you build a
package.  For packages built in the Fedora build system, they always get
a Vendor tag.  But if you build something locally, you probably don't
have a Vendor tag.  Account for this appropriately in the 'metadata'
inspection.

Also switch from using headerGetAsString() to headerGetString() to avoid
unnecessary memory allocation and leaking.