Add the 'ownership' inspection to librpminspect.
The ownership inspection enforces a handful of policies and reports
changes in file owner and group information.  It works for both single
builds as well as comparing two builds.  If you are just running
rpminspect against a single build, it skips the change reporting part of
the ownership inspection.  Here's what it does:

1) Iterates over every file in every package in the after build, but
skips source packages.

2) Checks to see if rpminspect has forbidden_owners set.  If so, check
if the file is owned by a forbidden owner.  If it is, report that as
RESULT_BAD.

3) Checks to see if rpminspect has forbidden_groups set.  If so, check
if the file is owned by a forbidden group.  If it is, report that as
RESULT_BAD.

4) Checks to see if the file is in a bin_path.  If it is, verifies the
file is owned by bin_owner and if not, reports that as RESULT_BAD.

5) If a file is in a bin_path and not group owned by bin_group, then it
performs some additional checks:
    a) Checks to see if CAP_SETUID is set on the file.
    b) If CAP_SETUID is set and the file is S_IXOTH, report as
       RESULT_BAD noting it is world executable.
    c) If CAP_SETUID is set and the file is S_IWGRP, report as
       RESULT_BAD requiring a Security Team review noting it is
       group writable.
    d) If CAP_SETUID is not set, report the file must be group
       owned by bin_group and is RESULT_BAD.

6) In the case of a before and after build, compare the owners and
groups of each file.  Report any changes as RESULT_VERIFY unless the
file is in a bin_path and the after build has changed to bin_owner and
bin_group.  In that case just report it as RESULT_INFO.
Iterate over each package in the after build:
    Iterate over each file