Ignore .pyc and .pyo Python bytecode files in 'changedfiles'
These are cached files that may or may not be in the package.  Since we
have the .py files, those are what really matters in terms of changing,
so ignore the bytecode files when looking for file differences.

Python bytecode files appear to begin with roughly the same 4 bytes.
The first byte looks to be a version number (like ABI or something) and
then you have 0x0D 0x0D 0x0A.  The changedfiles inspection just looks at
the latter 3 bytes to verify the .pyc or .pyo file is a Python bytecode
file.  Avoids having to chase version numbers.