Remove desktop_file_validate from rpminspect.conf
desktop_file_validate was the only command that could be set in the
configuration file.  The changedfiles inspection uses a lot of other
commands, but those are just macros in the source file.  Allowing
configuration file overrides for the commands may be a bit unnecessary
for now, but could be extended later on if we need that runtime
flexibility.

Consolidate the commands used in inspections to constants.h with a macro
name that ends with _CMD.  If I move these to a configuration file, I
might create a second file that is key=value just to keep the main file
restricted to test settings.