Add 'shells' setting to rpminspect.conf, default in constants.h
The shells setting is a space delimited list of shells that rpminspect
will use to perform syntax checking.  The shells listed must be actual
shells on the system, be listed by basename only, and support the -n
switch for syntax checking.  If not provided in the rpminspect.conf
file, rpminspect will default to the SHELLS list from constants.h.

Example list:

    sh ksh zsh csh tcsh rc bash