Add the 'shellsyntax' inspection.
This is another test inspired by rpminspect's ancestor.  Here's what it
does:

1) Shell scripts are run through their shell with the '-n' option
   and the output and return code is collected.
2) If the exit code is non-zero, the script is reported as invalid
   and marked RESULT_BAD.
3) If there is a before build, perform the same check on the script
   from the before build.
       a) If the script was bad before but good after, report
          RESULT_INFO that the script is now valid.
       b) If the script was good before but bad after, report
          RESULT_BAD that the script is now invalid.
4) In any case if the shell is 'bash' and the '-n' check returns
   non-zero and try it again with '-n -O extglob'.  If that passes,
   report the script is invalid with -n but valid with extglob and
   the script needs "shopt extglob" in it.

Some scripts report MIME types as "text/plain" and some are
"text/x-shellscript".  I stick with looking for "text/" and then trying
to read the #! line.

This inspection is not suitable for languages like Perl or Python
because those programs lack a syntax checker with a '-n' option.  Also,
Perl and Python programs tend to be more complicated than a shell
script.

Lastly, the list of shells this inspection works for is set in the
"shells" setting in rpminspect.conf.  The default list is:

    sh ksh zsh csh tcsh rc bash

You can modify this list or disable the inspection if you don't have
these shells available or just don't want to run the shellsyntax
inspection.  The shells listed in this setting must be the basename(3)
of the program and not the full path.  This means that rpminspect
expects them to be executable in the $PATH in its environment.