New integration test suite using the Python unittest module.
Move the existing tests to this new suite and make sure 'meson test'
uses the Python unittest module.  The main reason for moving the
integration test suite over to Python is to make use of the rpmfluff
Python module which can be used to create known good and bad builds for
the purposes of testing.