Fix a number of bugs in the disttag inspection.
First, return true when things are good and false for failures.  All of
the disttag failures are now RESULT_BAD.  Change the main driver to only
iterate over the packages first then over the files for the source RPMs.
Lastly, if we have scanned all packages available and never once saw an
SRPM file, report that as RESULT_BAD and tell the user.