Support runtime override of the result threshold (#47)
Add the -t option to let users specify a result code that sets the mark
to trigger a non-zero exit code from rpminspect.  By default this is
VERIFY meaning any inspection that hits a RESULT_VERIFY or higher result
will cause rpminspect to exit with EXIT_FAILURE.  Any code below that is
EXIT_SUCCESS.

You can pass any result code name (case-insensitive) to -t to set the
result threshold.  For example:

    rpminspect -t BAD
    rpminspect --threshold=INFO

The intent with the exit code of rpminspect is to indicate a passing or
failing inspection of a build.  Result codes of OK, INFO, or WAIVED are
not a failure.  VERIFY is the only odd one we have where regarding
VERIFY as a failure may seem somewhat pedantic.  VERIFY indicates
rpminspect does not know what the expected outcome should be for a
particular test, so it marks it as something the user should verify.
For BAD, rpminspect knows what the expected outcome is.