Do not mask upstream source changes (#51)
The upstream inspection was assuming that if the version of a package
changed between builds or the epoch changed then it could be implied
that an upstream source change was going to happen.  This behavior
was inherited from rpminspect's ancestor.  It does not seem correct to
mask this change, if present, because the tool should be reporting
changes regardless of the values of the version or epoch.  At least in
this inspection.