Simplify the upstream inspection more.
Everything I wrote yesterday for it was pointless.  I was making
it needlessly complicated when I didn't need to.  The version and
epoch don't matter in this inspection anymore, so it just needs
to report changes between source file peers and if any new ones
appear or existing ones disappear.