Match approved license tags with spaces in them (#53)
The license inspection was incorrectly handling whole license tag value
matching when spaces were involved.  Take "ASL 2.0" as an example.  It
was failing to show up as an approved license because the inspection code
was falling through to tokenizing the tag value and checking each component
against known license abbreviations.  The fall through was happening
because the whole tag check was using the license_name value which is the
JSON block identifier in the JSON license database in
/usr/share/rpminspect/licenses.  What it needed to do was check the whole
tag against the fedora_abbrev and spdx_abbrev.  It does that now and
"ASL 2.0" in the License: tag will show up as approved.

Added some test cases to account for this as well.

The tokenizing part happens because License tags in spec files can look
like:

    GPLv2+ and BSD and MIT
    BSD and (MPL or FooPL)

This is an established convention in Fedora and Fedora-like distributions.
The license inspection in rpminspect will go through this string and look
for individual license abbreviations.  It throws out the parens and the
boolean words.  Negation is not handled since it does not make sense to have
a License tag say something like "BSD and not GPL".  This code works fine
but the part missing was first checking to see if the entire License tag
value matches something in the approved list.