Support License tags with abbreviations with spaces and booleans.
Further exploration in to the license inspection revealed that boolean
expressions in the tag where one or more license abbreviation has spaces
would fail.  For example:

     GPLv3+ and ASL 2.0

Would fail to validate.  Adjust the code further to work with expressions
like this and add some test cases for it.  I have also confirmed it works
with some really long tags as well, like:

    MIT and Python and ASL 2.0 and BSD and ISC and LGPLv2 and MPLv2.0 and (ASL 2.0 or BSD)