Fix segfault when RPM path contains ..
When the path to the before or after RPM contained '..' in it,
rpminspect would previously segfault:

    $ rpminspect ../koji/resteasy-3.0.19-9.fc30.noarch.rpm resteasy-3.0.26-1.fc32.noarch.rpm
    *** Error extracting /var/tmp/rpminspect/local.6nixV7/before/../koji/resteasy-3.0.19-9.fc30.noarch.rpm: Path contains '..'
    Segmentation fault (core dumped)

This is undesirable. To prevent the segfault, fix error handling in
add_peer and get_rpm_info in order to detect when extract_rpm fails.

This has the side effect of showing a bogus second error message though:

    $ rpminspect ../koji/resteasy-3.0.19-9.fc30.noarch.rpm resteasy-3.0.26-1.fc32.noarch.rpm
    *** Error extracting /var/tmp/rpminspect/local.hcZD52/before/../koji/resteasy-3.0.19-9.fc30.noarch.rpm: Path contains '..'
    *** Error gathering build ../koji/resteasy-3.0.19-9.fc30.noarch.rpm: No such file or directory
    *** Failed to gather specified builds.

But at least it doesn't segfault any more.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>