Fix some fundamental problems with the peer detection code.
There were some issues here.  They popped up for me when working on
the test cases for the emptyrpm inspection.  I don't think these have
been widely reported yet because many users are just running
rpminspect across a single package or build.  But if you are doing
build comparisons, the peer detection code is key.

This is why I continue to work on the integration test suite.  It's
the only way to ensure what I've written is working the way I expect.