Fix a number of bugs in the desktop inspection.
A lot of changes here, but all for the better:

1) Fix the Exec= line check to account for desktop files that specify
arguments after the program name.  We don't need the arguments, just
the name of the program.

2) Fix the Exec= and Icon= check to work with symlinks.

3) Expand the Icon= check to honor desktop_icon_paths from the
rpminspect.conf file since icons may live in multiple locations on the
system.

4) Use strreplace() on the output from desktop-file-validate to
replace the fullpath with the localpath in the output.  The localpath
is what would be on an installed system while the fullpath is
localpath prefixed with the temporary working directory path for
rpminspect.