Fix the Icon file check in the desktop inspection.
This required a bit more reworking than I thought.  I renamed subtree
to allpkgtrees to better reflect what it is, then I use readdir() to
iterate over that and build icon search trees to feed to nftw() until
we might find the icon.  That should cover cases where the icon and
desktop file are in different subpackages but are otherwise provided
by the same SRPM.  It also checks for a basename match as well as a
match that ignores the extension.  What I am currently unsure about is
how to handle theme icon references, but I will deal with that when I
can find a better example.