Add support for the new >= 1.14.5 mandoc API (#15)
This one has been on the to do list for a while.  mandoc changed and
when we compile it in to a shared library, the API changed.  The API
changed completely between versions 1.14.4 and 1.14.5 which I think is
probably a side effect of this project being developed as part of a
BSD tree and then pulled out.

rpminspect currently supports < 1.14.5 and >= 1.14.5 APIs and will
detect that at build time.  Eventually the support for the older API
can be removed.