Use headerGetString() throughout librpminspect
librpm provides both headerGetString() and headerGetAsString() to get
header tag values as strings.  headerGetString() returns a pointer in
to the Header struct while headerGetAsString() strdup()'s the tag
value and returns a pointer to that.  If you use headerGetAsString(),
you need to free() the returned value.  Since librpminspect stores the
rpm Headers in the main program struct, we don't need copies of those
strings.  We also only ever use the strings for comparison and copying
in to formatted strings for the results struct.  We never modify the
tag value strings.

The use of headerGetAsString() was causing enormous memory leaks when
rpminspect compared very large packages (e.g., the kernel) which would
eventually lead to a crash.