Add get_nevra() to librpminspect to get RPMTAG_NEVRA
RPMTAG_NEVRA is an extension tag and lives in a different part of the
RPM header.  We have to use headerGet() for these types of tags and
then read the string that way.  Since there are two inspections using
this extension tag, add get_nevra() to rpm.c to make it easy.

NOTE:  Callers to get_nevra() must free the result.