Simplify the is_valid_license() code that concatenates tokens
Some licenses (e.g., ASL 2.0) are considered one token but when
walking the string in the code we break on spaces.  We consider 'and'
and 'or' to be the actual delimiters so any spaces occurring before
the next one should be considered part of the license token.

This commit does not change existing functionality, it just rearranges
things to make it more readable.