Run srpm and check targets before koji, pass name of tarball
When you run 'make koji', everything will happen that needs to happen
for it to generate a release tarball and spec file and then import new
builds in to dist-git and build them.  The 'make koji' target also
generates a rolling spec file compatible changelog file that is stored
in dist-git.  This is replaced in the spec file in dist-git so that
official builds have a rolling changelog.